+++ resources/views/librarian/author/index.blade.php
@@ -10,7 +10,7 @@
             <h2 class="page-title">Authors</h2>
         </div>
         <div class="col-auto ms-auto d-print-none">
-            <button class="btn btn-primary btn-5 d-none d-sm-inline-block" onclick="addAuthor()">
+            <button class="btn btn-primary btn-5" onclick="addAuthor()">
             <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-2">
                 <path d="M12 5l0 14"></path>
                 <path d="M5 12l14 0"></path>