+++ resources/views/librarian/patron/modal/patron_modal.blade.php
@@ -1,11 +1,10 @@
-{{-- Add Librarian Modal --}}
-<div class="modal modal-blur fade" id="add-librarian-modal" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
+<div class="modal modal-blur fade" id="add-patron-modal" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
     <div class="modal-dialog modal-lg modal-dialog-centered modal-dialog-scrollable" role="document">
         <div class="modal-content">
             
             <div class="modal-header">
-                <h5 class="modal-title">Add Librarian</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" onclick="resetAddLibrarian()"></button>
+                <h5 class="modal-title">Add Patron</h5>
+                <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
             </div>
 
             <div class="modal-body">
@@ -28,24 +27,12 @@
                     </svg>
                     </div>
                     <span>
-                        To access librarian account, Username and Password is the <span class="fw-bold">Librarian Code.</span> Please change your password once you get access to the account.
+                        To access patron account, Username and Password is the <span class="fw-bold">Patron Code.</span> Please change your password once you get access to the account.
                     </span>
                 </div>
-                <form method="POST" action="{{ route('admin.manage_librarians.store') }}" id="addLibrarianForm" enctype="multipart/form-data">
+                <form method="POST" action="{{ route('librarian.patron.store') }}" id="addPatronForm" enctype="multipart/form-data">
                 @csrf
                 <div class="mb-3">
-                    <label class="form-label">Librarian Code</label>
-                    <div class="row d-flex align-items-center">
-                        <div class="col-8">
-                            <input type="text" class="form-control" name="libraryCode" placeholder="Auto-Generated Librarian Code" id="libraryCode" maxlength="20" disabled>
-                        </div>
-                        <div class="col-4">
-                            <label class="form-check">
-                                <input class="form-check-input" type="checkbox" onclick="toggleAddCustomCode()" name="customCodeCheck" id="customCodeCheck">
-                                <span class="form-check-label">Custom Code</span>
-                            </label>
-                        </div>
-                    </div>
                     <div class="row d-flex align-items-center mt-3">
                         <div class="col-6">
                             <label class="form-label required">Last Name</label>
@@ -94,6 +81,20 @@
                             </select>
                         </div>
                         <div class="col-6">
+                            <label class="form-label required">Patron Type</label>
+                                <select class="form-select" name="patronType">
+                                    <option value="" selected disabled>Select Patron Type</option>
+                                    @if ($patronType)
+                                        @foreach ($patronType as $item)
+                                        <option value="{{ $item->id}}">{{$item->name}}</option>
+                                        @endforeach                                    
+                                    @endif
+                                </select>
+                            </select>
+                        </div>
+                    </div>
+                    <div class="row d-flex align-items-center mt-3">
+                        <div class="col-12">
                             <label class="form-label">Upload Profile Picture <small class="text-info fw-normal">(PNG and JPG only)</small></label>
                             <input type="file" class="form-control" accept="image/jpeg, image/png" name="profilePicture">
                         </div>
@@ -102,7 +103,7 @@
             </div>
 
             <div class="modal-footer">
-                <button type="submit" class="btn btn-primary" id="addLibrarianSubmit">
+                <button type="submit" class="btn btn-primary" id="addPatronSubmit">
                     Submit
                 </button>
             </div>
@@ -112,13 +113,15 @@
     </div>
 </div>
 
-<div class="modal modal-blur fade" id="edit-librarian-modal" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
+<div class="modal modal-blur fade" id="edit-patron-modal" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
     <div class="modal-dialog modal-lg modal-dialog-centered modal-dialog-scrollable" role="document">
         <div class="modal-content">
+            
             <div class="modal-header">
-                <h5 class="modal-title">Edit Librarian</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" onclick="resetEditLibrarian()"></button>
+                <h5 class="modal-title">Edit Patron</h5>
+                <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
             </div>
+
             <div class="modal-body">
                 <div class="alert alert-info" role="alert">
                     <div class="alert-icon">
@@ -130,29 +133,28 @@
                     </div>
                     Please fill up the required fields.
                 </div>
-
-                <form method="POST" action="{{ route('admin.manage_librarians.update') }}" id="editLibrarianForm" enctype="multipart/form-data">
+                <form method="POST" action="{{ route('librarian.patron.update') }}" id="editPatronForm" enctype="multipart/form-data">
                 @csrf
+                <input type="hidden" id="edit_user_id" name="edit_user_id">
                 <div class="mb-3">
-                    <input type="hidden" id="edit_user_id" name="edit_user_id">
                     <div class="row d-flex align-items-center mt-3">
                         <div class="col-6">
                             <label class="form-label required">Last Name</label>
-                            <input type="text" class="form-control" name="edit_lastName" id="edit_lastName" placeholder="Last Name" maxlength="20">
+                            <input type="text" class="form-control" id="edit_lastName" name="edit_lastName" placeholder="Last Name" maxlength="20">
                         </div>
                         <div class="col-6">
                             <label class="form-label required">First Name</label>
-                            <input type="text" class="form-control" name="edit_firstName" id="edit_firstName" placeholder="First Name" maxlength="20">
+                            <input type="text" class="form-control" id="edit_firstName" name="edit_firstName" placeholder="First Name" maxlength="20">
                         </div>
                     </div>
                     <div class="row d-flex align-items-center mt-3">
                         <div class="col-6">
                             <label class="form-label">Middle Name</label>
-                            <input type="text" class="form-control" name="edit_middleName" id="edit_middleName" placeholder="Middle Name" maxlength="20">
+                            <input type="text" class="form-control" id="edit_middleName" name="edit_middleName" placeholder="Middle Name" maxlength="20">
                         </div>
                         <div class="col-6">
                             <label class="form-label">Suffix</label>
-                            <select class="form-select" name="edit_suffix" id="edit_suffix">
+                            <select class="form-select" id="edit_suffix" name="edit_suffix">
                             <option value="">No suffix</option>
                             <option value="Jr.">Jr.</option>
                             <option value="Sr.">Sr.</option>
@@ -166,7 +168,7 @@
                     <div class="row d-flex align-items-center mt-3">
                         <div class="col-6">
                             <label class="form-label required">Contact Number</label>
-                            <input type="text" name="edit_contactNumber" id="edit_contactNumber" class="form-control" data-mask="09123456789" data-mask-visible="true" placeholder="09123456789" autocomplete="off" maxlength="11" inputmode="numeric" pattern="[0-9]*">
+                            <input type="text" id="edit_contactNumber" name="edit_contactNumber" class="form-control" data-mask="09123456789" data-mask-visible="true" placeholder="09123456789" autocomplete="off" maxlength="11" inputmode="numeric" pattern="[0-9]*">
                         </div>
                         <div class="col-6">
                             <label class="form-label required">Email Address</label>
@@ -176,10 +178,33 @@
                     <div class="row d-flex align-items-center mt-3">
                         <div class="col-6">
                             <label class="form-label required">Gender</label>
-                            <select class="form-select" name="edit_gender" id="edit_gender">
+                            <select class="form-select" id="edit_gender" name="edit_gender">
                             <option value="" selected disabled>Select Gender</option>
                             <option value="Male">Male</option>
                             <option value="Female">Female</option>
+                            </select>
+                        </div>
+                        <div class="col-6">
+                            <label class="form-label required">Patron Type</label>
+                                <select class="form-select" id="edit_patronType" name="edit_patronType">
+                                    <option value="" selected disabled>Select Patron Type</option>
+                                    @if ($patronType)
+                                        @foreach ($patronType as $item)
+                                        <option value="{{ $item->id}}">{{$item->name}}</option>
+                                        @endforeach                                    
+                                    @endif
+                                </select>
+                            </select>
+                        </div>
+                    </div>
+                    <div class="row d-flex align-items-center mt-3">
+                        <div class="col-6">
+                            <label class="form-label text-danger">Status <small class="text-info fw-normal">(Change at own risk)</small></label>
+                            <select class="form-select" id="edit_status" name="edit_status">
+                            <option value="" selected disabled>Status</option>
+                            <option value="active">Active</option>
+                            <option value="inactive">Inactive</option>
+                            <option value="suspended">Suspended</option>
                             </select>
                         </div>
                         <div class="col-6">
@@ -191,45 +216,12 @@
             </div>
 
             <div class="modal-footer">
-                <button type="submit" class="btn btn-primary" id="editLibrarianSubmit">
-                    Submit
+                <button type="submit" class="btn btn-primary" id="editPatronSubmit">
+                    Update
                 </button>
             </div>
                 
             </form>
-
-        </div>
-    </div>
-</div>
-
-<div class="modal modal-blur fade" id="view-inactive-librarian-modal" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
-    <div class="modal-dialog modal-xl modal-dialog-centered modal-dialog-scrollable" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title">Inactive Librarians</h5>
-            </div>
-            <div class="modal-body">
-                <table class="table datatable table-selectable table-vcenter text-nowrap table-responsive" id="inactiveLibrarianTable">
-                    <thead>
-                        <tr>
-                            <th>Librarian Code</th>
-                            <th>Profile</th>
-                            <th>Gender</th>
-                            <th>Action</th>
-                        </tr>
-                    </thead>
-                    <tbody id="inactiveLibrarianTbody">
-                    </tbody>
-                </table>
-            </div>
-
-            <div class="modal-footer">
-                <button type="button" class="btn btn-primary" class="btn-close" data-bs-dismiss="modal">
-                    Close
-                </button>
-            </div>
-                
-
         </div>
     </div>
 </div>