+++ resources/views/master/librarian_master.blade.php
@@ -138,8 +138,8 @@
                         <span class="nav-link-title"> Authors </span>
                       </a>
                     </li>
-                    <li class="nav-item">
-                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                    <li class="nav-item {{ request()->routeIs('librarian.patron.*') ? 'active' : '' }}">
+                      <a class="nav-link" href="{{ route('librarian.patron.index') }}">
                         <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
                             <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-user-cog">
                               <path stroke="none" d="M0 0h24v24H0z" fill="none" />